fix: expense amount를 Double -> Long으로 수정

+++ src/main/java/com/dnd/moddo/domain/expense/dto/request/ExpenseRequest.java
@@ -8,7 +8,7 @@ import com.dnd.moddo.domain.memberExpense.dto.request.MemberExpenseRequest;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 public record ExpenseRequest(
-	Double amount,
+	Long amount,
 	String content,
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
 	LocalDate date,
