+++ src/main/java/com/dnd/moddo/domain/memberExpense/dto/request/MemberExpenseRequest.java
@@ -1,4 +1,4 @@
 package com.dnd.moddo.domain.memberExpense.dto.request;
 
-public record MemberExpenseRequest(Long memberId, Double amount) {
+public record MemberExpenseRequest(Long memberId, Long amount) {
 }
